requirements.txt
python-dotenv==1.0.1
openai>=1.40.0
python-docx==1.2.0
tiktoken==0.11.0
langchain==0.3.27
langchain-community==0.3.27
faiss-cpu==1.8.0.post1
numpy==1.26.4
pytest==8.4.1
neo4j==5.28.2
streamlit==1.48.1
streamlit-webrtc==0.63.4
SpeechRecognition==3.14.3
pydub==0.25.1
soundfile==0.13.1
pandas==2.3.1
requests==2.32.4
whisper==1.1.10
httpx==0.28.1
streamlit-mic-recorder>=0.0.8